+++ appcli/app.go
@@ -20,6 +20,7 @@ func newBaseApp() *cli.App {
 	app.HideVersion = true
 	app.HideHelp = true
 	app.EnableBashCompletion = true
+	app.UseShortOptionHandling = true
 
 	app.Flags = append(app.Flags,
 		cli.BoolFlag{